+++ src/find-pairs/index.js
@@ -1,5 +1,6 @@
 const vscode = require('vscode');
 const log = require('../utils/log');
+const { CLOSURES, getEnabledClosures }= require('../utils/closures');
 const findBracketPairs = require('./brackets');
 const findTagPairs = require('./tags');
 
@@ -10,13 +11,29 @@ module.exports = () => {
 
   if (!activeEditor) return pairs;
 
-  let bracketPairs = findBracketPairs();
-  log(`Found ${bracketPairs.length}  bracket pairs!`, bracketPairs);
+  if (getEnabledClosures().includes(CLOSURES.PARENTHESES)) {
+    let parenthesesPairs = findBracketPairs(CLOSURES.PARENTHESES);
+    log(`Found ${parenthesesPairs.length} parentheses pairs! ()`, parenthesesPairs);
+    pairs = pairs.concat(parenthesesPairs);
+  }
+    
+  if (getEnabledClosures().includes(CLOSURES.CURLY_BRACKET)) {
+    let curlyBracketPairs = findBracketPairs(CLOSURES.CURLY_BRACKET);
+    log(`Found ${curlyBracketPairs.length} curly bracket pairs! {}`, curlyBracketPairs);
+    pairs = pairs.concat(curlyBracketPairs);
+  }
 
-  let tagPairs = findTagPairs();
-  log(`Found ${tagPairs.length}  tag pairs!`, tagPairs);
-
-  pairs = [].concat(bracketPairs).concat(tagPairs);
+  if (getEnabledClosures().includes(CLOSURES.SQUARE_BRACKET)) {
+    let squareBracketPairs = findBracketPairs(CLOSURES.SQUARE_BRACKET);
+    log(`Found ${squareBracketPairs.length} square bracket pairs! []`, squareBracketPairs);
+    pairs = pairs.concat(squareBracketPairs);
+  }
+  
+  if (getEnabledClosures().includes(CLOSURES.TAG)) {
+    let tagPairs = findTagPairs();
+    log(`Found ${tagPairs.length}  tag pairs!`, tagPairs);
+    pairs = pairs.concat(tagPairs);
+  } 
 
   // Remove pairs where opening and closing line are the same => there will never be a preview for them
   pairs = pairs.filter(pair => {