+++ src/utils/indexOf.js
@@ -1,11 +1,15 @@
+const { getEnabledClosingsRegex } = require('./closures');
+
 module.exports = {
   
-  regexIndexOf: (string, regex) => {
+  regexIndexOfClosing: (string) => {
+    const regex = getEnabledClosingsRegex();
     var match = string.match(regex);
     return match ? string.indexOf(match[0]) : -1;
   },
-
-  regexLastIndexOf: (string, regex) => {
+  
+  regexLastIndexOfClosing: (string) => {
+    const regex = getEnabledClosingsRegex();
     var match = string.match(regex);
     return match ? string.lastIndexOf(match[match.length - 1]) : -1;
   }