+++ test/files/test.m
@@ -1,0 +1,10 @@
+#import <Foundation/Foundation.h>
+ 
+int main () {
+
+   for( ; ; ) {
+      NSLog(@"This loop will run forever.\n");
+   }
+
+   return 0;
+}